Factory pattern added and salary modified to Double

+++ initialTest/src/main/resources/initial_test.sql
@@ -7,7 +7,7 @@ CREATE TABLE employees(
   id INT NOT NULL AUTO_INCREMENT PRIMARY KEY,
   name VARCHAR(255),
   email VARCHAR(255),
-  salary INT,
+  salary DOUBLE,
   birth_date DATE,
   department_id INT,
   FOREIGN KEY (department_id) REFERENCES departments (id)
